pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "celery[sqlalchemy]>=5.5.0rc3",
    "click>=8.1.8",
    "email-validator>=2.2.0",
    "emails>=0.6",
    "fastapi[standard]>=0.115.8",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.5",
    "psycopg[binary]>=3.2.4",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "rich>=13.9.4",
    "sqlalchemy[asyncio]>=2.0.37",
    "uvicorn-worker>=0.3.0",
]

[project.scripts]
appcli = "app.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "faker>=35.2.0",
    "httpx>=0.28.1",
    "mypy>=1.15.0",
    "polyfactory>=2.19.0",
    "pytest>=8.3.4",
    "ruff>=0.9.4",
]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]
exclude = ["venv", ".venv", "migrations"]

[tool.uv]
prerelease = "allow"

[tool.ruff]
exclude = ["migrations"]
